requirements.txt
psycopg2-binary==2.9.9
paramiko==3.4.0
python-telegram-bot==13.7